better file organization/naming; new api aux files

+++ api/remover_turma.php
@@ -4,11 +4,24 @@
 
   require_once __DIR__ . '/db_connect.php';
 
-  $queryTurma  = "DELETE FROM turma WHERE id = {$turma}";
-  $resultTurma = mysqli_query($connection, $queryTurma);
+  $queryAtividade  = "SELECT COUNT(*) as qtd FROM atividade WHERE id_turma = {$turma}";
+  $resultAtividade = mysqli_query($connection, $queryAtividade);
 
-  if ($resultTurma) {
-    echo True;
+  $qtd = mysqli_fetch_assoc($resultAtividade);
+
+  if ($qtd['qtd'] == 0) {
+    $queryUsuarioTurma  = "DELETE FROM usuario_turma WHERE id_turma = {$turma}";
+    $resultUsuarioTurma = mysqli_query($connection, $queryUsuarioTurma);
+
+    $queryTurma  = "DELETE FROM turma WHERE id = {$turma}";
+    $resultTurma = mysqli_query($connection, $queryTurma);
+
+    if ($resultUsuarioTurma && $resultTurma) {
+      echo True;
+    }
+    else {
+      echo False;
+    }
   }
   else {
     echo False;
